Preprocessing/requirements.txt
numpy==1.23.5
pickleshare==0.7.5
jsonschema==4.17.3
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
pandas==1.5.3

scikit-learn==1.2.2
pymongo==4.3.3
confluent-kafka==2.1.1
paho-mqtt==1.6.1
